frontend/requirements.txt
pygame==2.0.0
pygame-textinput==1.0.1
requests == 2.27.1
jose == 1.0.0
python-jose == 3.3.0
websocket-client == 1.3.2